# specs/openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce App API
  version: 1.0.0
  description: A fully-functioning e-commerce application REST API
paths:
...
  /users/{userId}/orders:
    get:
      operationId: getOrdersByUser
      summary: Get all Orders by User ID
      parameters:
      - $ref: '#/components/parameters/userId'
      responses:
        '200':
          $ref: '#/components/responses/ordersSuccess'
      tags:
        - User - Orders
    post:
      operationId: addOrderByUser
      summary: Add a new Order by User ID
      parameters:
      - $ref: '#/components/parameters/userId'
      requestBody:
        $ref: '#/components/requestBodies/NewOrderAndItems'
      responses:
        '201':
          $ref: '#/components/responses/orderSuccess'
      tags:
        - User - Orders
    delete:
      operationId: deleteAllOrdersByUser
      summary: Delete all Orders by User ID
      parameters:
      - $ref: '#/components/parameters/userId'
      responses:
        '204':
          $ref: '#/components/responses/deleteSuccess'
      tags:
        - User - Orders
...
  /orders/{orderId}:
    get:
      operationId: getOrderById
      summary: Get an Order by ID
      parameters:
      - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          $ref: '#/components/responses/orderSuccess'
      tags:
        - Orders
    put:
      operationId: updateOrderById
      summary: Update an Order by ID
      parameters:
      - $ref: '#/components/parameters/orderId'
      requestBody:
        $ref: '#/components/requestBodies/Order'
      responses:
        '200':
          $ref: '#/components/responses/orderSuccess'
      tags:
        - Orders
    delete:
      operationId: deleteOrderById
      summary: Delete an Order by ID
      parameters:
      - $ref: '#/components/parameters/orderId'
      responses:
        '204':
          $ref: '#/components/responses/deleteSuccess'
      tags:
        - Orders
  /orders/{orderId}/status:
    get:
      operationId: getOrderStatusById
      summary: Get an Order Status by ID
      parameters:
      - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          $ref: '#/components/responses/orderStatusSuccess'
      tags:
        - Orders
    put:
      operationId: updateOrderStatusById
      summary: Update an Order Status by ID
      parameters:
      - $ref: '#/components/parameters/orderId'
      requestBody:
        $ref: '#/components/requestBodies/OrderStatus'
      responses:
        '200':
          $ref: '#/components/responses/orderStatusSuccess'
      tags:
        - Orders
...
components:
  schemas:
    OrderStatus:
      type: string
      enum:
        - created
        - accepted
        - paid
        - posted
        - recieved
...
    NewOrder:
      type: object
      properties:
        created:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/OrderStatus'
    NewOrderItem:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        amount:
          type: integer
          format: int32
          minimum: 0
    NewOrderItems:
      type: array
      items:
        $ref: '#/components/schemas/NewOrderItem'
    NewOrderAndItems:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/NewOrder'
        orderItems:
          $ref: '#/components/schemas/NewOrderItems'
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        created:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/OrderStatus'
    Orders:
      type: array
      items:
        $ref: '#/components/schemas/Order'
  parameters:
    userId:
      name: userId
      in: path
      description: unique id for a User
      required: true
      schema:
        type: string
        format: uuid
...
    orderId:
      name: orderId
      in: path
      description: unique id for an Order
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    orderSuccess:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Order'
      description: Success
    ordersSuccess:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Orders'
      description: Success
...
    orderStatusSuccess:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OrderStatus'
      description: Success
    deleteSuccess:
      content:
        application/json:
          schema:
            type: string
          example: Delete operation was successful
      description: Delete operation was successful
...
  requestBodies:
...
    NewOrderAndItems:
      description: Information needed for creating a new Order object
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NewOrderAndItems'
...
    Order:
      description: An Order object
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Order'
      required: true
...
    OrderStatus:
      description: An Order Status
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OrderStatus'
      required: true
...
servers:
  - url: '{server}/api/v1'
    variables:
      server:
        default: http://127.0.0.1:4001